Correctly annotated

+++ 0x01-python_async_function/4-tasks.py
@@ -7,7 +7,7 @@ from typing import List
 task_wait_random = __import__('3-tasks').task_wait_random
 
 
-async def task_wait_n(n: int, max_delay: float) -> List[float]:
+async def task_wait_n(n: int, max_delay: int) -> List[float]:
     """
     Function calls the task_wait_random async function and returns
     a list of times
